--- _data/projects/archived-projects.ts
@@ -1,66 +1,0 @@
-// moved to `Ex-Projects`
-import { Projects } from "~/interfaces/project-types";
-
-export const archivedProjects: Projects[] = [
-  {
-    title: "The Hello World Project",
-    href: "https://github.com/skywalkerSam/the-hello-world-project",
-    description:
-      "Hello World for Every Programming Language possible! Hopefully.)",
-  },
-  {
-    title: "introductionToWebDevelopment",
-    href: "https://github.com/skywalkerSam/introductionToWebDevelopment",
-    description: "Introduction to Web Development, fr*",
-  },
-  {
-    title: "Project Cortana",
-    href: "https://github.com/skywalkerSam/Project-Cortana",
-    description: "Project Cortana. Inspired by Halo Infinite.)",
-  },
-  {
-    title: "JS 101",
-    href: "https://github.com/skywalkerSam/JS-101",
-    description: 'JavaScript 101 (Not, Java. "JavaScript".)',
-  },
-  {
-    title: "Python 101",
-    href: "https://github.com/skywalkerSam/Python-101",
-    description: "Introduction to Python 3",
-  },
-  {
-    title: "ML 101",
-    href: "https://github.com/skywalkerSam/ML-101",
-    description: "Machine Learning 101",
-  },
-  {
-    title: "Game Development",
-    href: "https://github.com/skywalkerSam/Game-Dev-101",
-    description: "Game Development 101",
-  },
-  {
-    title: "Guessing Game",
-    href: "https://github.com/skywalkerSam/Guessing-Game",
-    description: "A Terminal-based Guessing Game written in C++",
-  },
-  {
-    title: "Audio Generator",
-    href: "https://github.com/skywalkerSam/Audio-Generator",
-    description: "Text-to-Audio w/ Python 3",
-  },
-  {
-    title: "IT 101",
-    href: "https://github.com/skywalkerSam/IT-101",
-    description: "Getting Started with Information Technologies",
-  },
-  // {
-  //   title: "Linux 101",
-  //   href: "https://github.com/skywalkerSam/Linux-101",
-  //   description: "Linux 101",
-  // },
-  {
-    title: "Sidewalks",
-    href: "https://github.com/skywalkerSam/Sidewalks",
-    description: "Sidewalks...",
-  },
-];
