--- _data/projects/ex-projects.ts
@@ -1,132 +1,0 @@
-import { Projects } from "~/interfaces/project-types";
-
-export const exProjects: Projects[] = [
-  {
-    title: "The Hello World Project",
-    href: "https://github.com/skywalkerSam/the-hello-world-project",
-    description:
-      "Hello World for Every Programming Language possible! Hopefully.)",
-  },
-  {
-    title: "introductionToWebDevelopment",
-    href: "https://github.com/skywalkerSam/introductionToWebDevelopment",
-    description: "Introduction to Web Development, fr*",
-  },
-  {
-    title: "JS 101",
-    href: "https://github.com/skywalkerSam/JS-101",
-    description: 'JavaScript 101 (Not Java, "JavaScript".)',
-  },
-  {
-    title: "Project Cortana",
-    href: "https://github.com/skywalkerSam/Project-Cortana",
-    description: "Project Cortana. Inspired by Halo Infinite.)",
-  },
-  {
-    title: "Python 101",
-    href: "https://github.com/skywalkerSam/Python-101",
-    description: "Introduction to Python 3",
-  },
-  {
-    title: "Hello React!",
-    href: "https://github.com/skywalkerSam/hello-react",
-    description: "The Hello World of React.",
-  },
-  {
-    title: "ML 101",
-    href: "https://github.com/skywalkerSam/ML-101",
-    description: "Machine Learning 101",
-  },
-  {
-    title: "Dashboard Demo",
-    href: "https://github.com/skywalkerSam/dashboard-demo",
-    description: "Dashboard Demo w/ NEXT.js + Vercel.",
-  },
-  {
-    title: "House Of Cards",
-    href: "https://github.com/skywalkerSam/houseOfCards",
-    description:
-      "A simple game where you'd waste your time matching two cards at a time.)",
-  },
-  {
-    title: "Tic Tac Toe",
-    href: "https://github.com/skywalkerSam/tictactoe",
-    description: "Simple Tic Tac Toe.)",
-  },
-  {
-    title: "Face Detection",
-    href: "https://github.com/skywalkerSam/faceDetection",
-    description: "Basic Face Detection w/ Clarifai.",
-  },
-  {
-    title: "Dall-E Wrapper",
-    href: "https://github.com/skywalkerSam/dalleWrapper",
-    description:
-      "Mimicking Dall-E-like Responses w/ Randomized Preset Images.)",
-  },
-  {
-    title: "Face Detection Demo",
-    href: "https://github.com/skywalkerSam/face-detection-vite",
-    description: "Face Detection Demo built w/ Vite + React.",
-  },
-  {
-    title: "Animal Farm",
-    href: "https://github.com/skywalkerSam/animalSearch",
-    description: "POC for Searching Things.",
-  },
-  {
-    title: "Timer",
-    href: "https://github.com/skywalkerSam/timer",
-    description: "A Simple Timer made w/ React.",
-  },
-  {
-    title: "Clock",
-    href: "https://github.com/skywalkerSam/clock",
-    description: "A Simple Clock made w/ React.",
-  },
-  {
-    title: "To Do",
-    href: "https://github.com/skywalkerSam/todo",
-    description: "A Simple To Do made w/ React.",
-  },
-  {
-    title: "Workspace",
-    href: "https://github.com/skywalkerSam/workspace",
-    description: "A Simple Workspace made w/ React.",
-  },
-  {
-    title: "Robofriends",
-    href: "https://github.com/skywalkerSam/robofriends",
-    description: "A Collection of Robot Cards made w/ React.",
-  },
-  {
-    title: "Audio Generator",
-    href: "https://github.com/skywalkerSam/Audio-Generator",
-    description: "Text-to-Audio w/ Python 3",
-  },
-  {
-    title: "Guessing Game",
-    href: "https://github.com/skywalkerSam/Guessing-Game",
-    description: "A Terminal-based Guessing Game written in C++",
-  },
-  {
-    title: "Game Development",
-    href: "https://github.com/skywalkerSam/Game-Dev-101",
-    description: "Game Development 101",
-  },
-  {
-    title: "IT 101",
-    href: "https://github.com/skywalkerSam/IT-101",
-    description: "Getting Started with Information Technologies",
-  },
-  // {
-  //   title: "Linux 101",
-  //   href: "https://github.com/skywalkerSam/Linux-101",
-  //   description: "Linux 101",
-  // },
-  {
-    title: "Sidewalks",
-    href: "https://github.com/skywalkerSam/Sidewalks",
-    description: "Sidewalks...",
-  },
-];
