--- _data/projects/projects.ts
@@ -1,14 +1,0 @@
-import { Projects } from "~/interfaces/project-types";
-
-export const majorProjects: Projects[] = [
-  {
-    title: "Project: Stardate",
-    href: "https://github.com/skywalkerSam/stardate",
-    description: "A better way of dating things (at least for me.)",
-  },
-  {
-    title: "Next Gallery",
-    href: "https://github.com/skywalkerSam/next-gallery",
-    description: "Personalized Cloud-based Image Gallery w/ Next.js + Vercel.",
-  },
-];
