--- src/interfaces/project-types.ts
@@ -1,5 +1,0 @@
-export type Projects = {
-  title: string;
-  href: string;
-  description: string;
-};
